modularized api startup to allow swagger schema generation
