+++ api/src/main.ts
@@ -9,50 +9,14 @@ import {
 } from './common/constants/swagger.constants';
 import { PrismaClientExceptionFilter } from 'nestjs-prisma';
 import { ConfigService } from '@nestjs/config';
+import { bootstrap, setupSwagger } from './bootstrap';
 
-async function bootstrap() {
-  const app = await NestFactory.create(AppModule, {
-    logger: ['error', 'warn', 'log', 'debug', 'verbose'],
-  });
-
-  const configService = app.get(ConfigService);
-
-  const { httpAdapter } = app.get(HttpAdapterHost);
-  app.useGlobalFilters(new PrismaClientExceptionFilter(httpAdapter));
-
-  const corsOrigin = configService.get<string>('API_CORS_ORIGIN');
-
-  app.enableCors({
-    origin: corsOrigin ?? 'https://easymotion.devlocal',
-    methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
-    credentials: true, // Allows cookies or auth headers
-  });
-
-  app.useGlobalPipes(
-    new ValidationPipe({
-      transform: true,
-      whitelist: true,
-      forbidNonWhitelisted: false,
-    }),
-  );
-
-  app.useGlobalInterceptors(new ClassSerializerInterceptor(app.get(Reflector)));
-
+async function start() {
+  const app = await bootstrap();
   if (process.env.NODE_ENV !== 'production') {
-    const config = new DocumentBuilder()
-      .setTitle(API_TITLE)
-      .setDescription(API_DESCRIPTION)
-      .setVersion(API_VERSION)
-      .build();
-
-    const document = SwaggerModule.createDocument(app, config);
-
-    SwaggerModule.setup('swagger', app, document, {
-      jsonDocumentUrl: 'swagger/json',
-    });
+    setupSwagger(app);
   }
-
-  await app.listen(process.env.PORT ?? 3000);
+  await app.listen(3000);
 }
 
-bootstrap();
+start();
